+++ packages/app-builder-lib/src/codeSign/windowsCodeSign.ts
@@ -6,7 +6,7 @@ import { computeToolEnv, ToolInfo } from "../util/bundledTool"
 import { rename } from "fs-extra"
 import * as os from "os"
 import * as path from "path"
-import { resolveFunction } from "../config/resolve"
+import { resolveFunction } from "../util/resolve"
 import { isUseSystemSigncode } from "../util/flags"
 import { VmManager } from "../vm/vm"
 import { WinPackager } from "../winPackager"
