+++ packages/app-builder-lib/src/util/NodeModuleCopyHelper.ts
@@ -4,7 +4,7 @@ import { lstat, readdir, lstatSync } from "fs-extra"
 import * as path from "path"
 import { excludedNames, FileMatcher } from "../fileMatcher"
 import { Packager } from "../packager"
-import { resolveFunction } from "../config/resolve"
+import { resolveFunction } from "./resolve"
 import { FileCopyHelper } from "./AppFileWalker"
 import { NodeModuleInfo } from "./packageDependencies"
 import { realpathSync } from "fs"
