+++ packages/app-builder-lib/src/util/resolve.ts
